requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
dataclasses-json==0.6.7
distro==1.9.0
faiss-cpu==1.11.0
fastapi==0.103.1
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.32.4
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-cli==0.0.21
langchain-community==0.3.24
langchain-core==0.3.63
langchain-openai==0.3.19
langchain-text-splitters==0.3.8
langserve==0.0.41
langsmith==0.1.147
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
mysql-connector-python==9.3.0
networkx==3.5
numpy==1.26.4
openai==1.82.1
orjson==3.10.18
packaging==23.2
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
propcache==0.3.1
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyMuPDF==1.26.0
pypdfium2==4.30.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==1.8.2
starlette==0.27.0
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomlkit==0.12.5
torch==2.7.0
tqdm==4.67.1
transformers==4.52.4
typer==0.9.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.23.2
yarl==1.20.0
zstandard==0.23.0
